% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childes_types.R
\docType{data}
\name{childes_types}
\alias{childes_types}
\title{child_types}
\format{A data frame with 19,375 rows and 2 columns:
\describe{
\item{\code{type}}{a character denoting a word type}
\item{\code{n_type}}{number of occurences of word across all 14 corpora}
}}
\source{
Data is from the CHILDES Database (\url{https://childes.talkbank.org}) and is accessed using the \code{childesr} package (\url{https://github.com/langcog/childesr}).
}
\usage{
childes_types
}
\description{
A data frame containing the counts of word types in \code{childes_tokens}
across all 14 corpora.
}
\keyword{datasets}
