% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childes_utterances.R
\docType{data}
\name{childes_utterances}
\alias{childes_utterances}
\title{child_utterances}
\format{A data frame with 237,577 rows representing individual utterances and 7 columns:
\describe{
\item{\code{collection}}{a factor denoting the CHILDES collection of the corpus}
\item{\code{corpus}}{a factor denoting corpus name}
\item{\code{child}}{a character string giving the child's name}
\item{\code{sex}}{a factor giving the child's sex}
\item{\code{age}}{a double giving the child's age}
\item{\code{transcript_id}}{integer denoting the transcript ID of an utterance}
\item{\code{utterance}}{a character string giving an utterance}
}}
\source{
Data is from the CHILDES Database (\url{https://childes.talkbank.org}) and is accessed using the \code{childesr} package (\url{https://github.com/langcog/childesr}).
}
\usage{
childes_utterances
}
\description{
A data frame containing utterances of children from 14 corpora in the
CHILDES database (2018.1).
}
\keyword{datasets}
